refactor: renames <Login/> page-component, test files and improves its compositions
