+++ template/src/pages/login/__tests__/Login.test.tsx
@@ -1,32 +1,38 @@
 import userEvent from '@testing-library/user-event';
-import Login from '../index';
+import Login from '../Login';
 import { renderWithRouter, screen, waitFor } from '../../../utils/test';
 import { makeServer } from '../../../server';
 
-let server: any;
-
-beforeEach(() => {
-  server = makeServer();
-});
-
-afterEach(() => {
-  server.shutdown();
-});
-
 describe('<Login />', () => {
-  it('should render the login form UI', () => {
+  // eslint-disable-next-line
+  let server: any;
+
+  beforeEach(() => {
+    server = makeServer();
+  });
+
+  afterEach(() => {
+    server.shutdown();
+  });
+
+  it('should render the login form UI', async () => {
     renderWithRouter(<Login />);
 
-    expect(
-      screen.getByText(/fill the fields below to login/i)
-    ).toBeInTheDocument();
-    expect(screen.getByLabelText('email')).toBeInTheDocument();
-    expect(screen.getByLabelText('password')).toBeInTheDocument();
-    expect(screen.getByRole('button', { name: /login/i })).toBeInTheDocument();
+    await waitFor(() => {
+      expect(
+        screen.getByText(/fill the fields below to login/i)
+      ).toBeInTheDocument();
+      expect(screen.getByLabelText('email')).toBeInTheDocument();
+      expect(screen.getByLabelText('password')).toBeInTheDocument();
+      expect(
+        screen.getByRole('button', { name: /login/i })
+      ).toBeInTheDocument();
+    });
   });
 
   it('should render error messages when inputs are invalid', async () => {
     renderWithRouter(<Login />);
+
     userEvent.type(screen.getByLabelText('email'), 'not valid');
     userEvent.type(screen.getByLabelText('password'), 'small{enter}');
 
@@ -35,7 +41,7 @@ describe('<Login />', () => {
         screen.getByText(/insert a valid email address/i)
       ).toBeInTheDocument();
       expect(
-        screen.getByText(/The password must have at leat 7 characters./i)
+        screen.getByText(/The password must have at least 7 characters./i)
       ).toBeInTheDocument();
     });
   });
@@ -50,6 +56,7 @@ describe('<Login />', () => {
     userEvent.type(screen.getByLabelText('email'), 'user@t10.digital');
     userEvent.type(screen.getByLabelText('password'), '111111111111');
     userEvent.click(screen.getByRole('button', { name: /login/i }));
+
     await waitFor(() => {
       expect(
         screen.getByText(/Email or password are incorrect\. Try again\./i)
